Another method
Here is the another method to predict that the person has dibetes or not. In this method I have used one code for predictive system. We can make predictive systems like this using the same method, this mehthod can be used as a blueprint for making predictive systems.Predictive system if the way by which we can predict the result of the model which we have trained.